chore(idex): move DC-related commands from ratos.py to ratos_dual_carriage_extras.py
